+++ evolve/template/constants.go
@@ -1,0 +1,20 @@
+package template
+
+const (
+	ServerAddCommandsWithStartCmdOptions = "server.AddCommandsWithStartCmdOptions"
+
+	RollkitV0XStartHandler = "rollserv.StartHandler"
+	EvolveV1XStartHandler  = "abciserver.StartHandler"
+	evolveV1MigrateCmd     = "abciserver.MigrateToRollkitCmd()"
+)
+
+const (
+	GoExecPackage = "github.com/evstack/ev-abci"
+	GoExecVersion = "v0.3.0"
+
+	EvNodePackage = "github.com/evstack/ev-node"
+	EvNodeVersion = "v1.0.0-beta.2"
+
+	EvNodeDaCmd     = "github.com/evstack/ev-node/da/cmd/local-da"
+	EvNodeDaVersion = "v1.0.0-beta.1"
+)